+++ src/main/java/crm/common/model/Customer.java
@@ -10,8 +10,8 @@ import crm.core.repository.persistence.entity.relation.FetchMode;
 @Entity(tableName = "Customer")
 public class Customer {
     @Key
-    @Column(name = "CustomerID", type = "INT")
-    private Integer customerID;
+    @Column(name = "CustomerID", type = "BIGINT")
+    private Long customerID;
 
     @Column(name = "CustomerName", length = 150, nullable = false)
     private String customerName;
@@ -31,11 +31,11 @@ public class Customer {
     @ManyToOne(joinColumn = "Username", fetch = FetchMode.EAGER)
     private LazyReference<Account> account;
 
-    public Integer getCustomerID() {
+    public Long getCustomerID() {
         return customerID;
     }
 
-    public void setCustomerID(Integer customerID) {
+    public void setCustomerID(Long customerID) {
         this.customerID = customerID;
     }
