+++ src/main/java/crm/common/model/Feature.java
@@ -7,8 +7,8 @@ import crm.core.repository.persistence.annotation.Key;
 @Entity(tableName = "Feature")
 public class Feature {
     @Key
-    @Column(name = "FeatureID", type = "INT")
-    private Integer featureID;
+    @Column(name = "FeatureID", type = "BIGINT")
+    private Long featureID;
 
     @Column(name = "FeatureURL", length = 255, nullable = false)
     private String featureURL;
@@ -16,10 +16,27 @@ public class Feature {
     @Column(name = "Description", length = 255)
     private String description;
 
-    public Integer getFeatureID() { return featureID; }
-    public void setFeatureID(Integer featureID) { this.featureID = featureID; }
-    public String getFeatureURL() { return featureURL; }
-    public void setFeatureURL(String featureURL) { this.featureURL = featureURL; }
-    public String getDescription() { return description; }
-    public void setDescription(String description) { this.description = description; }
+    public Long getFeatureID() {
+        return featureID;
+    }
+
+    public void setFeatureID(Long featureID) {
+        this.featureID = featureID;
+    }
+
+    public String getFeatureURL() {
+        return featureURL;
+    }
+
+    public void setFeatureURL(String featureURL) {
+        this.featureURL = featureURL;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
 }
