+++ src/main/java/crm/common/model/InventoryItem.java
@@ -7,23 +7,23 @@ import crm.core.repository.persistence.entity.relation.FetchMode;
 @Entity(tableName = "InventoryItem")
 public class InventoryItem {
     @Key
-    @Column(name = "Item_ID", type = "INT")
-    private Integer itemId;
+    @Column(name = "Item_ID", type = "BIGINT")
+    private Long itemId;
 
     @Column(name = "SerialNumber", length = 255, nullable = false, unique = true)
     private String serialNumber;
 
-    @Column(name = "ProductID", type = "INT", nullable = false)
-    private Integer productID;
+    @Column(name = "ProductID", type = "BIGINT", nullable = false)
+    private Long productID;
 
     @ManyToOne(joinColumn = "ProductID", fetch = FetchMode.EAGER)
     private LazyReference<Product> product;
 
-    public Integer getItemId() {
+    public Long getItemId() {
         return itemId;
     }
 
-    public void setItemId(Integer itemId) {
+    public void setItemId(Long itemId) {
         this.itemId = itemId;
     }
 
@@ -35,11 +35,11 @@ public class InventoryItem {
         this.serialNumber = serialNumber;
     }
 
-    public Integer getProductID() {
+    public Long getProductID() {
         return productID;
     }
 
-    public void setProductID(Integer productID) {
+    public void setProductID(Long productID) {
         this.productID = productID;
     }
