+++ src/main/java/crm/common/model/ProductSpecification.java
@@ -7,14 +7,14 @@ import crm.core.repository.persistence.entity.relation.FetchMode;
 @Entity(tableName = "ProductSpecification")
 public class ProductSpecification {
     @Key
-    @Column(name = "ProductSpecificationID", type = "INT")
-    private Integer productSpecificationID;
+    @Column(name = "ProductSpecificationID", type = "BIGINT")
+    private Long productSpecificationID;
 
-    @Column(name = "ProductID", type = "INT", nullable = false)
-    private Integer productID;
+    @Column(name = "ProductID", type = "BIGINT", nullable = false)
+    private Long productID;
 
-    @Column(name = "SpecificationID", type = "INT", nullable = false)
-    private Integer specificationID;
+    @Column(name = "SpecificationID", type = "BIGINT", nullable = false)
+    private Long specificationID;
 
     @ManyToOne(joinColumn = "ProductID", fetch = FetchMode.EAGER)
     private LazyReference<Product> product;
@@ -22,27 +22,27 @@ public class ProductSpecification {
     @ManyToOne(joinColumn = "SpecificationID", fetch = FetchMode.EAGER)
     private LazyReference<Specification> specification;
 
-    public Integer getProductSpecificationID() {
+    public Long getProductSpecificationID() {
         return productSpecificationID;
     }
 
-    public void setProductSpecificationID(Integer productSpecificationID) {
+    public void setProductSpecificationID(Long productSpecificationID) {
         this.productSpecificationID = productSpecificationID;
     }
 
-    public Integer getProductID() {
+    public Long getProductID() {
         return productID;
     }
 
-    public void setProductID(Integer productID) {
+    public void setProductID(Long productID) {
         this.productID = productID;
     }
 
-    public Integer getSpecificationID() {
+    public Long getSpecificationID() {
         return specificationID;
     }
 
-    public void setSpecificationID(Integer specificationID) {
+    public void setSpecificationID(Long specificationID) {
         this.specificationID = specificationID;
     }
