+++ src/main/java/crm/common/model/ProductWarehouse.java
@@ -10,18 +10,18 @@ import crm.core.repository.persistence.entity.relation.FetchMode;
 @Entity(tableName = "ProductWarehouse")
 public class ProductWarehouse {
     @Key
-    @Column(name = "ProductWarehouseID", type = "INT")
-    private Integer productWarehouseID;
+    @Column(name = "ProductWarehouseID", type = "BIGINT")
+    private Long productWarehouseID;
 
     @Column(name = "ProductStatus", length = 20)
     @Convert(converter = ProductStatusConverter.class)
     private ProductStatus productStatus;
 
-    @Column(name = "WarehouseID", type = "INT", nullable = false)
-    private Integer warehouseID;
+    @Column(name = "WarehouseID", type = "BIGINT", nullable = false)
+    private Long warehouseID;
 
-    @Column(name = "ItemID", type = "INT", nullable = false)
-    private Integer itemID;
+    @Column(name = "ItemID", type = "BIGINT", nullable = false)
+    private Long itemID;
 
     @ManyToOne(joinColumn = "WarehouseID", fetch = FetchMode.EAGER)
     private LazyReference<Warehouse> warehouse;
@@ -34,11 +34,11 @@ public class ProductWarehouse {
     @OneToOne(mappedBy = "productWarehouse", joinColumn = "ProductWarehouseID", fetch = FetchMode.LAZY)
     private LazyReference<ProductExported> productExported;
 
-    public Integer getProductWarehouseID() {
+    public Long getProductWarehouseID() {
         return productWarehouseID;
     }
 
-    public void setProductWarehouseID(Integer productWarehouseID) {
+    public void setProductWarehouseID(Long productWarehouseID) {
         this.productWarehouseID = productWarehouseID;
     }
 
@@ -50,19 +50,19 @@ public class ProductWarehouse {
         this.productStatus = productStatus;
     }
 
-    public Integer getWarehouseID() {
+    public Long getWarehouseID() {
         return warehouseID;
     }
 
-    public void setWarehouseID(Integer warehouseID) {
+    public void setWarehouseID(Long warehouseID) {
         this.warehouseID = warehouseID;
     }
 
-    public Integer getItemID() {
+    public Long getItemID() {
         return itemID;
     }
 
-    public void setItemID(Integer itemID) {
+    public void setItemID(Long itemID) {
         this.itemID = itemID;
     }
