+++ src/main/java/crm/common/model/Specification.java
@@ -9,8 +9,8 @@ import java.util.List;
 @Entity(tableName = "Specification")
 public class Specification {
     @Key
-    @Column(name = "SpecificationID", type = "INT")
-    private Integer specificationID;
+    @Column(name = "SpecificationID", type = "BIGINT")
+    private Long specificationID;
 
     @Column(name = "SpecificationName", length = 100, nullable = false)
     private String specificationName;
@@ -18,8 +18,8 @@ public class Specification {
     @Column(name = "SpecificationValue", length = 255, nullable = false)
     private String specificationValue;
 
-    @Column(name = "SpecificationTypeID", type = "INT", nullable = false)
-    private Integer specificationTypeID;
+    @Column(name = "SpecificationTypeID", type = "BIGINT", nullable = false)
+    private Long specificationTypeID;
 
     @ManyToOne(joinColumn = "SpecificationTypeID", fetch = FetchMode.EAGER)
     private LazyReference<SpecificationType> specificationType;
@@ -27,11 +27,11 @@ public class Specification {
     @OneToMany(mappedBy = "specificationID", joinColumn = "SpecificationID", fetch = FetchMode.LAZY)
     private List<ProductSpecification> productSpecifications;
 
-    public Integer getSpecificationID() {
+    public Long getSpecificationID() {
         return specificationID;
     }
 
-    public void setSpecificationID(Integer specificationID) {
+    public void setSpecificationID(Long specificationID) {
         this.specificationID = specificationID;
     }
 
@@ -51,11 +51,11 @@ public class Specification {
         this.specificationValue = specificationValue;
     }
 
-    public Integer getSpecificationTypeID() {
+    public Long getSpecificationTypeID() {
         return specificationTypeID;
     }
 
-    public void setSpecificationTypeID(Integer specificationTypeID) {
+    public void setSpecificationTypeID(Long specificationTypeID) {
         this.specificationTypeID = specificationTypeID;
     }
