+++ src/main/java/crm/common/model/SpecificationType.java
@@ -9,14 +9,14 @@ import java.util.List;
 @Entity(tableName = "SpecificationType")
 public class SpecificationType {
     @Key
-    @Column(name = "SpecificationTypeID", type = "INT")
-    private Integer specificationTypeID;
+    @Column(name = "SpecificationTypeID", type = "BIGINT")
+    private Long specificationTypeID;
 
     @Column(name = "SpecificationTypeName", length = 100, nullable = false)
     private String specificationTypeName;
 
-    @Column(name = "TypeID", type = "INT", nullable = false)
-    private Integer typeID;
+    @Column(name = "TypeID", type = "BIGINT", nullable = false)
+    private Long typeID;
 
     @ManyToOne(joinColumn = "TypeID", fetch = FetchMode.EAGER)
     private LazyReference<Type> type;
@@ -24,11 +24,11 @@ public class SpecificationType {
     @OneToMany(mappedBy = "specificationTypeID", joinColumn = "SpecificationTypeID", fetch = FetchMode.LAZY)
     private List<Specification> specifications;
 
-    public Integer getSpecificationTypeID() {
+    public Long getSpecificationTypeID() {
         return specificationTypeID;
     }
 
-    public void setSpecificationTypeID(Integer specificationTypeID) {
+    public void setSpecificationTypeID(Long specificationTypeID) {
         this.specificationTypeID = specificationTypeID;
     }
 
@@ -40,11 +40,11 @@ public class SpecificationType {
         this.specificationTypeName = specificationTypeName;
     }
 
-    public Integer getTypeID() {
+    public Long getTypeID() {
         return typeID;
     }
 
-    public void setTypeID(Integer typeID) {
+    public void setTypeID(Long typeID) {
         this.typeID = typeID;
     }
