+++ src/main/java/crm/common/repository/CategoryRepository.java
@@ -3,7 +3,7 @@ package crm.common.repository;
 import crm.common.model.Category;
 import crm.core.repository.persistence.repository.AbstractRepository;
 
-public class CategoryRepository extends AbstractRepository<Category, Integer> {
+public class CategoryRepository extends AbstractRepository<Category, Long> {
     public CategoryRepository() {
         super(Category.class);
     }
