+++ src/main/java/crm/common/repository/FeedbackRepository.java
@@ -3,7 +3,7 @@ package crm.common.repository;
 import crm.common.model.Feedback;
 import crm.core.repository.persistence.repository.AbstractRepository;
 
-public class FeedbackRepository extends AbstractRepository<Feedback, Integer> {
+public class FeedbackRepository extends AbstractRepository<Feedback, Long> {
     public FeedbackRepository() {
         super(Feedback.class);
     }
