+++ src/main/java/crm/common/repository/InventoryItemRepository.java
@@ -3,7 +3,7 @@ package crm.common.repository;
 import crm.common.model.InventoryItem;
 import crm.core.repository.persistence.repository.AbstractRepository;
 
-public class InventoryItemRepository extends AbstractRepository<InventoryItem, Integer> {
+public class InventoryItemRepository extends AbstractRepository<InventoryItem, Long> {
     public InventoryItemRepository() {
         super(InventoryItem.class);
     }
