+++ src/main/java/crm/common/repository/ProductContractRepository.java
@@ -3,7 +3,7 @@ package crm.common.repository;
 import crm.common.model.ProductContract;
 import crm.core.repository.persistence.repository.AbstractRepository;
 
-public class ProductContractRepository extends AbstractRepository<ProductContract, Integer> {
+public class ProductContractRepository extends AbstractRepository<ProductContract, Long> {
     public ProductContractRepository() {
         super(ProductContract.class);
     }
