+++ src/main/java/crm/common/repository/ProductTransactionRepository.java
@@ -3,7 +3,7 @@ package crm.common.repository;
 import crm.common.model.ProductTransaction;
 import crm.core.repository.persistence.repository.AbstractRepository;
 
-public class ProductTransactionRepository extends AbstractRepository<ProductTransaction, Integer> {
+public class ProductTransactionRepository extends AbstractRepository<ProductTransaction, Long> {
     public ProductTransactionRepository() {
         super(ProductTransaction.class);
     }
